+++ packages/client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @liteforge/client
 
+## 8.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.9.0
+  - @liteforge/query@6.0.0
+
 ## 7.0.0
 
 ### Patch Changes