+++ packages/flow/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @liteforge/flow
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @liteforge/runtime@0.9.0
+
 ## 2.0.0
 
 ### Patch Changes